MVVM pattern implemented
An MVVM pattern has been implemented.  The search button on the app now
has basic functionality and if a valid ticker is entered will display
the name and last ticker price of the stock on the main page.

A view model has been created for a property on the service layer.

There is a bug where if a user hits the top left "back" button without
first choosing a stock at least once, the app crashes.

Error handling and validation still needs to be implemented.  The app
will crash if a non valid ticker is selected.