Calculate Page Built
The Calculate Page has been built, with appropriate properties binded to
the viewmodel.  The search page is performing slower now for some
reason, and a thread warning is given in the IDE.  The calculate method
now needs creation and the layout needs to be customized.  Validation
and error handling still haven't been implemented.